% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_trending_topics_analysis.R
\name{analyze_trends}
\alias{analyze_trends}
\title{Analyze keyword trends, visualize, and save to CSV}
\usage{
analyze_trends(tfidf_df)
}
\arguments{
\item{tfidf_df}{A data frame containing TF-IDF scores for keywords.}
}
\value{
None
}
\description{
This function analyzes keyword trends based on TF-IDF scores, visualizes the top keywords, and saves the results to CSV and PNG files.
}
\details{
This function calculates the mean TF-IDF scores for each keyword in the input data frame, identifies the top 10 keywords with the highest TF-IDF scores, and saves them to a CSV file named 'top_keywords_tfidf.csv'. It also creates a bar plot visualizing the top keywords and saves it as a PNG file named 'tfidf_plot.png'. Finally, it prints the plot.
}
\examples{
\dontrun{
analyze_trends(tfidf_scores)
}
}
