% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_trending_topics_analysis.R
\name{calculate_tfidf}
\alias{calculate_tfidf}
\title{Calculate TF-IDF (Term Frequency-Inverse Document Frequency)}
\usage{
calculate_tfidf(data)
}
\arguments{
\item{data}{A data frame containing text data with a 'title' column.}
}
\value{
A data frame containing TF-IDF scores.
}
\description{
This function calculates TF-IDF scores for text data.
}
\details{
This function preprocesses the 'title' column of the input data by converting text to lowercase, removing punctuation and stopwords, and then calculates TF-IDF scores using the \code{tm} package.
}
\examples{
\dontrun{
tfidf_scores <- calculate_tfidf(data)
}
}
